Fix number of arguments issue for direct dispatch and rename module to direct
